+++ Actions/Extreme/TowersDealLessDamage.cs
@@ -30,5 +30,5 @@ public class TowersDealLessDamage : KeystrokeAction
     }
 
     /// <inheritdoc />
-    protected override float Weight => 25;
+    protected override int Weight => 25;
 }
