+++ Actions/Extreme/TowersHaveLessPierce.cs
@@ -31,5 +31,5 @@ public class TowersHaveLessPierce : KeystrokeAction
     }
 
     /// <inheritdoc />
-    protected override float Weight => 25;
+    protected override int Weight => 25;
 }
