+++ Actions/Normal/FasterBloons.cs
@@ -15,5 +15,5 @@ public class FasterBloons : KeystrokeAction
     }
 
     /// <inheritdoc />
-    protected override float Weight => 400;
+    protected override int Weight => 400;
 }
