+++ Actions/Normal/MonkeysGrow.cs
@@ -29,5 +29,5 @@ public class MonkeysGrow : KeystrokeAction
 
 
     /// <inheritdoc />
-    protected override float Weight => 400;
+    protected override int Weight => 400;
 }
