+++ Actions/Normal/TowerRangeDecreased.cs
@@ -31,5 +31,5 @@ public class TowerRangeDecreased : KeystrokeAction
     }
 
     /// <inheritdoc />
-    protected override float Weight => 400;
+    protected override int Weight => 400;
 }
