+++ Actions/Rare/MoreBloonsEachRound.cs
@@ -20,5 +20,5 @@ public class MoreBloonsEachRound : KeystrokeAction
     }
 
     /// <inheritdoc />
-    protected override float Weight => 100;
+    protected override int Weight => 100;
 }
